requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==5.1.1
idna==3.3
py==1.11.0
pyTelegramBotAPI==4.4.0
python-dotenv==0.20.0
requests==2.27.1
retry==0.9.2
urllib3==1.26.9